Be smarter about encoding/decoding [#18]
We were running into errors sometimes when encoding/decoding values taht
had already been compressed.  This patch updates the VariablePrimitive
class to be smarter about doing both.

The type of the value instance is checked against bytes or strings or
unicode (using the six compatability library) to ensure that we always
end up rendering a sequence of bytes and those bytes are either
pre-supplied or stringified and utf-8 encoded.

When decoding we assume the value is encoded in utf-8 and move along if
it is not (as in the case of gzip'ed nested message sets)